Move DateAndTime component to General
