+++ components/General/index.tsx
@@ -4,6 +4,46 @@ import Tooltip from 'antd/lib/tooltip';
 import React from 'react';
 import styled from 'styled-components';
 
+export const DateAndTime = styled(({ className, date, time }) => {
+    return (
+        <div className={className}>
+            {date}
+            <div className="secondary">{time}</div>
+        </div>
+    );
+})`
+    font-size: 1rem;
+    padding: 0.5rem 1rem;
+
+    ${(props) => {
+        if (props.borderTop === true) {
+            return 'border-top: 1px solid #002886;';
+        }
+    }};
+
+    ${(props) => {
+        if (props.borderRight === true) {
+            return 'border-right: 1px solid #002886;';
+        }
+    }};
+
+    ${(props) => {
+        if (props.borderBottom === true) {
+            return 'border-bottom: 1px solid #002886;';
+        }
+    }};
+
+    ${(props) => {
+        if (props.borderLeft === true) {
+            return 'border-left: 1px solid #002886;';
+        }
+    }};
+
+    .secondary {
+        color: ${(props) => props.color as string};
+    }
+`;
+
 export const Box: React.FC<{ className?: string }> = styled.div`
     border: 1px solid #0c3586;
     display: flex;
