+++ components/Portfolio/index.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import React from 'react';
 
 export const LeftPanel = styled.div`
     width: 25%;
@@ -147,46 +146,6 @@ export const TableCell = styled.td`
 
 export const SecondaryCell = styled.div`
     color: #005ea4;
-`;
-
-export const DateAndTime = styled(({ className, date, time }) => {
-    return (
-        <div className={className}>
-            {date}
-            <div className="secondary">{time}</div>
-        </div>
-    );
-})`
-    font-size: 1rem;
-    padding: 0.5rem 1rem;
-
-    ${(props) => {
-        if (props.borderTop === true) {
-            return 'border-top: 1px solid #002886;';
-        }
-    }};
-
-    ${(props) => {
-        if (props.borderRight === true) {
-            return 'border-right: 1px solid #002886;';
-        }
-    }};
-
-    ${(props) => {
-        if (props.borderBottom === true) {
-            return 'border-bottom: 1px solid #002886;';
-        }
-    }};
-
-    ${(props) => {
-        if (props.borderLeft === true) {
-            return 'border-left: 1px solid #002886;';
-        }
-    }};
-
-    .secondary {
-        color: ${(props) => props.color as string};
-    }
 `;
 
 export const StatusIndicator = styled.div`
